refactor: separate record_service for controller from odos_fast_record #43
